+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.6] - 2025-04-09
+
+### Added
+
+- Adds support for the "<audio>" tag using the "Goatee.Audio()" class.
+
 ## [0.2.5] - 2025-04-09
 
 ### Fixed